% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investment_path.R
\name{investment_path}
\alias{investment_path}
\title{Read investment candidates with their characteristics}
\usage{
investment_path(directory_path, path_solver, display = TRUE, report = TRUE,
  clean = TRUE, parallel = TRUE)
}
\arguments{
\item{directory_path}{Character containing the directory path
\code{antaresRead::setSimulationPath}}
}
\value{
|
Returns a list containing the different investment candidates. |
}
\description{
\code{investment_path}
}
\details{
is a function which reads the investment path, i.e the selected years for the study,
of the expansion planning problem and their characteristics. The information on
the candidates is stored in the file antaresStudyPath/user/expansion/studies.ini.
}
