% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_studies.R
\name{read_studies}
\alias{read_studies}
\title{Read studies with their characteristics}
\usage{
read_studies(studies_file_name)
}
\arguments{
\item{studies}{list of the paths of the studies returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Returns a list containing the different paths of the studies
WARNING: / may have to be changed to backslash
}
\description{
\code{read_studies} is a function which stores the links of the studies in a vector indexed on the selected years.
The information is stored in the file: antaresStudyPath/user/expansion/studies.ini.
}
