% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_studies_OLD.R
\name{read_studies_OLD}
\alias{read_studies_OLD}
\title{Read studies with their characteristics}
\usage{
read_studies_OLD(studies_file_name)
}
\arguments{
\item{studies}{list of the paths of the studies returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Returns a data frame containing the different paths of the studies:
WARNING: / may have to be changed to backslash 
simulated_years    paths
1            2020   D:/Users/manthube/Documents/Antares_work
2            2040   D:/Users/manthube/Documents
3            2050   D:/Users/manthube
}
\description{
\code{read_studies} is a function which stores the links of the studies in a vector indexed on the selected years.
The information is stored in the file: antaresStudyPath/user/expansion/studies.ini.
}
